# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Contactless cashless event payment system API (Colombia market)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/initiate:
    post:
      tags: [TopUps]
      operationId: initiateDigitalTopUp
      summary: Initiate a digital top-up via Nequi or PSE (Wompi)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiateDigitalTopUpBody'
      responses:
        '201':
          description: Payment intent created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WompiPaymentIntentResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          description: Payment gateway error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /payments/{id}/status:
    get:
      tags: [TopUps]
      operationId: getDigitalTopUpStatus
      summary: Poll status of a digital top-up payment intent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment intent status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WompiPaymentIntentStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/webhook:
    post:
      tags: [TopUps]
      operationId: handleWompiWebhook
      summary: Wompi webhook receiver (HMAC-validated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook received.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
...
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Not authenticated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
...
    WompiPaymentIntentResult:
      type: object
      required: [intentId, status, paymentMethod]
      properties:
        intentId:
          type: string
        status:
          type: string
          enum: [pending, success, failed]
        paymentMethod:
          type: string
          enum: [nequi, pse]
        wompiTransactionId:
          type: [string, 'null']
        redirectUrl:
          type: [string, 'null']
    WompiPaymentIntentStatus:
      type: object
      required: [intentId, status]
      properties:
        intentId:
          type: string
        status:
          type: string
          enum: [pending, success, failed]
        topUpId:
          type: [string, 'null']
    InitiateDigitalTopUpBody:
      type: object
      required: [braceletUid, amount, paymentMethod]
      properties:
        braceletUid:
          type: string
          minLength: 1
        amount:
          type: integer
          minimum: 1000
        paymentMethod:
          type: string
          enum: [nequi, pse]
        phoneNumber:
          type: string
        bankCode:
          type: string
...